Thêm chức năng review phía client và admin

+++ nhom10/src/main/java/com/haui/repository/UserRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long>, JpaSpecificationExecutor<User> {
+public interface UserRepository extends JpaRepository<User, Integer>, JpaSpecificationExecutor<User> {
     Page<User> findAll(Pageable pageable);
 
     Page<User> findAll(Specification<User> spec, Pageable pageable);
